new cocoa disease model added

+++ app/src/main/java/com/neotelemetrixgdscunand/kamekapp/data/remote/dto/WeatherForecastOverviewDto.kt
@@ -14,7 +14,7 @@ data class WeatherForecastOverviewDto(
     val windVelocity: Float? = null,
 
     @field:SerializedName("rainfall")
-    val rainfall: Int? = null,
+    val rainfall: Float? = null,
 
     @field:SerializedName("humidity")
     val humidity: Int? = null,
